+++ src/DDTrace/Integrations/Guzzle/V5/GuzzleIntegration.php
@@ -3,13 +3,13 @@
 namespace DDTrace\Integrations\Guzzle\V5;
 
 use DDTrace\Configuration;
-use DDTrace\Formats;
-use DDTrace\Tags;
-use DDTrace\Span;
-use DDTrace\Types;
+use DDTrace\Format;
 use DDTrace\GlobalTracer;
 use DDTrace\Http\Urls;
 use DDTrace\Integrations\Integration;
+use DDTrace\Span;
+use DDTrace\Tag;
+use DDTrace\Type;
 use GuzzleHttp\Message\ResponseInterface;
 
 class GuzzleIntegration extends Integration
@@ -35,8 +35,8 @@ class GuzzleIntegration extends Integration
     public static function setDefaultTags(Span $span, $method)
     {
         parent::setDefaultTags($span, $method);
-        $span->setTag(Tags\SPAN_TYPE, Types\HTTP_CLIENT);
-        $span->setTag(Tags\SERVICE_NAME, 'guzzle');
+        $span->setTag(Tag::SPAN_TYPE, Type::HTTP_CLIENT);
+        $span->setTag(Tag::SERVICE_NAME, 'guzzle');
     }
 
     /**
@@ -58,7 +58,7 @@ class GuzzleIntegration extends Integration
 
         $context = $span->getContext();
         $tracer = GlobalTracer::get();
-        $tracer->inject($context, Formats\HTTP_HEADERS, $headers);
+        $tracer->inject($context, Format::HTTP_HEADERS, $headers);
         $request->setHeaders($headers);
     }
 }