+++ tests/Unit/TracerTest.php
@@ -6,7 +6,7 @@ use DDTrace\Contracts\Tracer as TracerInterface;
 use DDTrace\OpenTracer\Tracer as OpenTracer;
 use DDTrace\Sampling\PrioritySampling;
 use DDTrace\SpanContext;
-use DDTrace\Tags;
+use DDTrace\Tag;
 use DDTrace\Tests\DebugTransport;
 use DDTrace\Time;
 use DDTrace\Tracer;
@@ -42,7 +42,7 @@ final class TracerTest extends Framework\TestCase
      */
     public function testCreateSpanSuccessWithExpectedValues(TracerInterface $tracer)
     {
-        $startTime = Time\now();
+        $startTime = Time::now();
         $span = $tracer->startSpan(self::OPERATION_NAME, [
             'tags' => [
                 self::TAG_KEY => self::TAG_VALUE
@@ -63,14 +63,14 @@ final class TracerTest extends Framework\TestCase
             'child_of' => $context,
         ]);
         $this->assertEquals($context->getSpanId(), $span->getParentId());
-        $this->assertNull($span->getTag(Tags\PID));
+        $this->assertNull($span->getTag(Tag::PID));
     }
 
     public function testStartSpanAsRootWithPid()
     {
         $tracer = new Tracer(new NoopTransport());
         $span = $tracer->startSpan(self::OPERATION_NAME);
-        $this->assertEquals(getmypid(), $span->getTag(Tags\PID));
+        $this->assertEquals(getmypid(), $span->getTag(Tag::PID));
     }
 
     public function testStartActiveSpan()
@@ -85,7 +85,7 @@ final class TracerTest extends Framework\TestCase
         $tracer = new Tracer(new NoopTransport());
         $parentScope = $tracer->startActiveSpan(self::OPERATION_NAME);
         $parentSpan = $parentScope->getSpan();
-        $parentSpan->setTag(Tags\SERVICE_NAME, 'parent_service');
+        $parentSpan->setTag(Tag::SERVICE_NAME, 'parent_service');
         $childScope = $tracer->startActiveSpan(self::ANOTHER_OPERATION_NAME);
         $this->assertEquals($childScope, $tracer->getScopeManager()->getActive());
         $this->assertEquals($parentScope->getSpan()->getSpanId(), $childScope->getSpan()->getParentId());