+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file - [read more
 - Move integrations tests to tests root folder #200
 - Removal of external dependencies to support auto-instrumentation #206
 - Allow testing of multiple library versions #203
+- Migration from namespace based constants to class based constants for tags, formats and types #207
 - Downgrade of phpunit to 4.* in order to prepare for php 5.4 #208
 
 ### Fixed